requirements.txt
# HSH Predictor — Dépendances Python

# ML
xgboost==2.1.4
scikit-learn==1.4.2
imbalanced-learn==0.12.3
joblib==1.3.2

# Data
numpy==1.26.4
pandas==2.2.1

# Base de données
psycopg2-binary==2.9.9

# API
requests==2.31.0
anthropic==0.25.0
httpx<0.28
fastapi==0.115.6
uvicorn==0.32.1

# Utilitaires
python-dotenv==1.0.1
